% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagefy.R
\name{packagefy}
\alias{packagefy}
\title{Install the packages required to make the \code{snaverse}run}
\usage{
packagefy(install = "all")
}
\arguments{
\item{install}{if "all" is selected, all the required packages will be installed.
If "CRAN" is selected, only the packages from the official R repository
will be installed. If "GitHun" is selected, only the packages hosted on GitHub
by developers will be installed.}
}
\value{
message for successful/unsuccessful installation.
}
\description{
Install the packages required to make the \code{snaverse}run
}
\details{
The \code{snaverse} is a suite of packages to learn and perform social network analysis.
The function \code{packagefy} installs the packages required to make the packages
inside the \code{snaverse} work properly.
There are three installation options
}
\examples{
\dontrun{
packagefy(install = "all")
}
}
