% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudiofy.R
\name{rstudiofy}
\alias{rstudiofy}
\title{Set up your Rstudio in the \code{SNA4DS}way}
\usage{
rstudiofy(setup = "SNA4DS")
}
\arguments{
\item{setup}{set to "SNA4DS". If any other option is specified, the function
will return a message suggesting to explore alternative Rstudio set up options
using the package \code{rstudio.prefs} and the function \code{use_rstudio_prefs}.}
}
\value{
A message that informs about succesfull/unsuccesful installation.
}
\description{
Set up your Rstudio in the \code{SNA4DS}way
}
\details{
This function automatically sets up your Rstudio IDE in the optimized way needed for the
SNA4DS course.
There is one installation option
}
\examples{
\dontrun{
rstudiofy(setup = "SNA4DS")
}
}
