% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaversify.R
\name{snaversify}
\alias{snaversify}
\title{Install the packages belonging to the \code{snaverse}}
\usage{
snaversify(install = "snaverse")
}
\arguments{
\item{install}{if "snaverse" is selected, three packages will be installed.
If "sna4tutti" is selected, only the package containing tutorials to learn sna
will be installed. If "snafun" is selected, only the package that provides functions
to perform sna will be installed. If "SNA4DSData" is selected, only the package
hosting network data set to perform sna will be installed.}
}
\value{
message for successful/unsuccessful installation.
}
\description{
Install the packages belonging to the \code{snaverse}
}
\details{
The \code{snaverse} is a suite of packages to learn and perform social network analysis.
The function \code{snaversify} installs the packages in the suite.
There are four installation options
}
\examples{
\dontrun{
snaversify(install = "snaverse")
}
}
